requirements.txt
addict==2.4.0
basicsr==1.4.2
cityscapesscripts==2.2.2
dlib==19.24.2
einops==0.7.0
face_alignment==1.3.5
ftfy==6.1.3
lmdb==1.4.1
matplotlib==3.7.3
mc==0.0
numpy==1.24.3
onnx==1.15.0
opencv_python==4.8.1.78
packaging==23.2
parrots==0.1.7
Pillow==9.3.0
prettytable==3.9.0
PyYAML==6.0.1
regex==2023.8.8
scipy==1.11.4
skimage==0.0
tensorboardX==2.6.2.2
tifffile==2023.7.10
timm==0.9.12
torch==1.12.0+cu102
torch_tb_profiler==0.4.3
torchvision==0.13.0
tqdm==4.66.1
turbojpeg==0.0.2
yapf==0.40.2